Move enum-to-class to its own utility header
